Refactored forms and views into package.
