+++ silver_coin/budget/urls.py
@@ -1,8 +1,7 @@
 from django.urls import path
 
-from .views import DashboardView, BudgetCreateView
+from .views import DashboardView
 
 urlpatterns = [
     path("dashboard/", DashboardView.as_view(), name="dashboard"),
-    path("budget_create/", BudgetCreateView.as_view(), name="budget_create"),
 ]
