+++ silver_coin/budget/views/__init__.py
@@ -1,0 +1,1 @@
+from .dashboard import DashboardView
